requirements.txt
aiohttp==2.0.7
Fiona==1.7.6
GDAL==2.1.3
mercantile==0.10.0
numpy==1.12.1
pyrr==0.9.2
olefile==0.44
Pillow==4.1.0
PyOpenGL==3.1.1a1
PyOpenGL-accelerate==3.1.1a1
pyproj==1.9.5.1
rasterio==1.0a5
Shapely==1.5.17
wxPython==4.0.0a1
git+https://github.com/imageio/imageio.git
matplotlib==2.0.2
netCDF4==1.2.9
requests==2.18.2
triangle==20170429
